plant-backend/requirements.txt
Flask>=2.3.0
flask-cors>=4.0.0
psycopg2-binary>=2.9.0
gunicorn>=20.1.0